Update README.md
In this SQL & Power BI project, I analyzed the data of a pizza store and used Power BI for visualization.
First, I downloaded the Excel file of the data online and used MS Excel for data cleaning. 
Secondly, I used SQL to perform various calculations. 
In the third step, I used POWER BI to represent the findings in the form of an interactive dashboard.